requirements.txt
pylast>=5.2.0
python-dotenv>=1.0.0
colorama>=0.4.6